@thinkeridea 2020-01-06
赞，这非常的棒，我想它不仅更快也会更安全,`unsafe` 包使用单个语句会避免指针丢失的情况，之前的方案我并没有遇到这个问题，但是理论上它是存在的，新的方案完美解决了这个问题👍